src/backend/requirements.txt
paddleocr
opencv-python
pyserial